Create a separate `SystemStatusReport` type just for SSR.
