+++ Networking/Networking/Model/SystemStatusDetails/SystemStatus+DropinMustUsePlugin.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public extension SystemStatus {
+public extension SystemStatusReport {
     /// Detail about drop-in / must-use plugin, which has minimal details compared to SystemPlugin
     ///
     struct DropinMustUsePlugin: Decodable {
